Change Opacity value type float->double
Change Opacity value type float->double ，because the internal API use the double type

+++ sources/engine/Stride.Games/Desktop/GameWindowWinforms.cs
@@ -274,7 +274,7 @@ namespace Stride.Games
         }
 
         /// <inheritdoc />
-        public override float Opacity { get; set; }
+        public override double Opacity { get; set; }
 
         /// <inheritdoc />
         public override Color BackgroundColor { get; set; }
