+++ sources/engine/Stride.Games/SDL/GameWindowSDL.cs
@@ -245,7 +245,7 @@ namespace Stride.Games
         }
 
         /// <inheritdoc />
-        public override float Opacity { get; set; }
+        public override double Opacity { get; set; }
 
         /// <inheritdoc />
         public override Color BackgroundColor { get; set; }
